Extract credential validation logic and use StatefulGuard::once
